more cleanup and shedding non-relevant tests
