+++ conda/gateways/logging.py
@@ -2,15 +2,16 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 from functools import partial
+import logging
 from logging import CRITICAL, DEBUG, ERROR, Filter, Formatter, INFO, StreamHandler, WARN, getLogger
 import re
 import sys
 
-from .. import CondaError, TRACE
+from .. import CondaError
 from .._vendor.auxlib.logz import NullHandler
 from ..common.io import attach_stderr_handler
 
-log = getLogger(__name__)
+TRACE = 5  # TRACE LOG LEVEL
 VERBOSITY_LEVELS = (WARN, INFO, DEBUG, TRACE)
 
 
@@ -86,3 +87,13 @@ def set_verbosity(verbosity_level):
     except IndexError:
         raise CondaError("Invalid verbosity level: %(verbosity_level)s",
                          verbosity_level=verbosity_level)
+
+
+def trace(self, message, *args, **kwargs):
+    if self.isEnabledFor(TRACE):
+        self._log(TRACE, message, args, **kwargs)
+
+
+logging.addLevelName(TRACE, "TRACE")
+logging.Logger.trace = trace
+initialize_logging()
